style(cache): reorder tools.py
